Refactor, components should stay in /components
