+++ app/blog/page.tsx
@@ -1,5 +1,5 @@
 import { Suspense } from "react";
-import BlogList from "./BlogList";
+import BlogList from "../components/BlogList";
 
 export default function Blog() {
   return (
